Move assoc type translation
